requirements.txt
asgiref==3.5.2
Django==4.0.5
mysqlclient==2.1.0
psycopg-binary==3.0.15
psycopg2-binary==2.9.3
sqlparse==0.4.2
